feat: create WhiteRabbit burner plugin
