+++ packages/burner-plugin/package.json
@@ -1,32 +1,32 @@
 {
-  "name": "@whiterabbitjs/wallet",
+  "name": "@whiterabbitjs/burner-plugin",
   "version": "1.0.0",
   "private": true,
   "license": "MIT",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "dependencies": {
     "@burner-wallet/assets": "^1.0.0",
     "@burner-wallet/core": "^1.0.3",
     "@burner-wallet/modern-ui": "^1.0.0",
-    "@burner-wallet/ui-core": "^1.0.7",
     "@whiterabbitjs/client": "^1.0.0",
-    "@whiterabbitjs/burner-plugin": "^1.0.0",
+    "styled-components": "^5.0.1"
+  },
+  "peerDependencies": {
     "react": "^16.13.1",
     "react-dom": "^16.13.1"
   },
   "devDependencies": {
     "@burner-wallet/types": "^1.0.6",
-    "@types/node": "^13.9.3",
     "@types/react": "^16.9.25",
     "@types/react-dom": "^16.9.5",
-    "react-scripts": "^3.4.1",
+    "copyfiles": "^2.2.0",
     "typescript": "^3.8.3"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "start-local": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "css": "copyfiles -u 1 \"src/**/*.css\" dist/",
+    "build": "tsc && npm run css",
+    "start-local": "tsc -w"
   },
   "browserslist": {
     "production": [
