+++ packages/burner-plugin/src/index.ts
@@ -1,0 +1,3 @@
+import WhiteRabbitPlugin from './WhiteRabbitPlugin';
+
+export default WhiteRabbitPlugin;
