+++ packages/burner-plugin/src/ui/MoviePage.module.css
@@ -1,0 +1,3 @@
+.moviePage {
+  padding: 0;
+}
