--- packages/wallet/src/components/Page/ErrorBoundary.tsx
@@ -1,20 +1,0 @@
-import React, { Component } from 'react';
-
-export default class ErrorBoundary extends Component<{}, { error?: any }> {
-  state = { error: null };
-
-  componentDidCatch(error: any) {
-    this.setState({ error });
-  }
-
-  render() {
-    if (this.state.error) {
-      return (
-        <pre>
-          Error - {`${this.state.error}`}
-        </pre>
-      );
-    }
-    return this.props.children;
-  }
-}
