--- packages/wallet/src/components/Page/index.tsx
@@ -1,2 +1,0 @@
-import Page from './Page';
-export default Page;
