--- packages/wallet/src/pages/MoviePage/index.tsx
@@ -1,2 +1,0 @@
-import MoviePage from './MoviePage';
-export default MoviePage;
